+++ apps/web/src/lib/queries/index.ts
@@ -8,5 +8,4 @@ export * from './partyInvites';
 export * from './partyMembers';
 export * from './scenes';
 export * from './stripe';
-export * from './useGetSceneTimestampsQuery';
 export * from './users';
